feat(papers): wire full-text extraction into the council, closing README gap #3
Adds a system-level FindingExtractor pipeline that runs after ACQUISITION:
Unpaywall lookup -> PDF fetch (new FullTextFetcher, deliberately bypassing
ToolGateway per CLAUDE.md 16) -> page parsing -> model extraction -> citation
location check -> persistence to studies/findings/citation_anchors, the first
production writers those tables have had. A located quote now reaches
Level A and emits a StudyFinding event, letting sql_projector's existing
DERIVED_FROM edge logic fire for the first time. Everything else (no OA
copy, unparsable PDF, unlocatable quote, budget exhaustion) is recorded as
an unfilled slot rather than faked, per CLAUDE.md 7.3.

Threaded through CouncilOrchestrator and apps/worker/jobs.py/main.py
alongside the existing acquirer. Also fixes two real bugs surfaced only by
running the pipeline against a real Postgres session in the integration
suite (unit tests used a fake session that can't enforce FK constraints):
finding_extraction.py never persisted the SourceVersion row that
studies.source_version_id points to, and its writes relied on SQLAlchemy
inferring insert order across five FK-linked tables that declare no
relationship() to each other -- so it now adds and flushes each row in
explicit dependency order. Also fixes a pre-existing string mismatch in
evidence/gate.py Stage 3 (compared against "FINDING" instead of the real
"StudyFinding" enum value) that would have silently skipped the check.

Adds an OpenAI-Chat-Completions-compatible model gateway and HTTP tool
gateway (packages/models/openai_compatible.py, packages/tools/http_gateway.py)
so the model/tool layer can target DeepSeek/LongCat/a domestic relay via
config rather than an Anthropic-specific implementation, with regression
tests guarding against ever falling back to this session's own
ANTHROPIC_AUTH_TOKEN/ANTHROPIC_BASE_URL as a vendor credential.

Verified: mypy --strict (219 files), ruff, 323 unit/golden/e2e tests, and
the full 180-test Docker-backed integration suite, all green.

+++ packages/models/contracts.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from decimal import Decimal
 from enum import StrEnum
 from typing import Protocol
 from uuid import UUID
@@ -39,7 +40,7 @@ class ModelResult(ContractModel):
     payload: FrozenDict[str, object]
     input_tokens: int
     output_tokens: int
-    cost_usd: int
+    cost_usd: Decimal
     latency_ms: int
     retries: int
     schema_status: SchemaStatus
